Beginnings of a Genes mechanic
I created the Gene class and IGenetic interface to map out the functionality of gene expression an inheritance in the simulation.

At the moment, I'm thinking that each gene will have some specific trait associated with it, that will currently be represented by an Enum (probably not the best tbh). Additionally, it will contain some value that contains the "genetic data" for an organism to start of with, and/or pass on to its offspring.

For inheritance, I figured implementing an IGenetic interface would be sufficient. This will allow both animals and plants to implement inheritance in the simulation.